refactor: create dto folders
