+++ src/main/java/com/ivocorrea/investmanager/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.ivocorrea.investmanager.controller;
 
-import com.ivocorrea.investmanager.dto.PutUserDto;
+import com.ivocorrea.investmanager.dto.user.PutUserDto;
 import com.ivocorrea.investmanager.entity.User;
 import com.ivocorrea.investmanager.service.UserService;
 import org.springframework.http.ResponseEntity;
