+++ src/main/java/com/ivocorrea/investmanager/dto/asset/PutAssetDTO.java
@@ -1,4 +1,4 @@
-package com.ivocorrea.investmanager.dto;
+package com.ivocorrea.investmanager.dto.asset;
 
 public record PutAssetDTO(int quantity, double currentPrice) {
 }
