+++ src/main/java/com/ivocorrea/investmanager/dto/login/LoginRequestDTO.java
@@ -1,4 +1,4 @@
-package com.ivocorrea.investmanager.dto;
+package com.ivocorrea.investmanager.dto.login;
 
 public record LoginRequestDTO(String email, String password) {
 }
