+++ src/main/java/com/ivocorrea/investmanager/dto/login/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ivocorrea.investmanager.dto;
+package com.ivocorrea.investmanager.dto.login;
 
 public record LoginResponseDTO(String acessToken, String refreshToken) {
 }
