+++ src/main/java/com/ivocorrea/investmanager/dto/portfolio/CreatePortfolioDTO.java
@@ -1,4 +1,4 @@
-package com.ivocorrea.investmanager.dto;
+package com.ivocorrea.investmanager.dto.portfolio;
 
 import java.util.UUID;
