+++ src/main/java/com/ivocorrea/investmanager/dto/user/CreateUserDto.java
@@ -1,4 +1,4 @@
-package com.ivocorrea.investmanager.dto;
+package com.ivocorrea.investmanager.dto.user;
 
 public record CreateUserDto(String username, String email, String password) {
 }
