+++ src/main/java/com/ivocorrea/investmanager/dto/user/PutUserDto.java
@@ -1,4 +1,4 @@
-package com.ivocorrea.investmanager.dto;
+package com.ivocorrea.investmanager.dto.user;
 
 public record PutUserDto(String username, String email) {
